Added consistency throughout the code. (#11)

+++ database/migrations/2018_07_01_123200_cart.php
@@ -32,7 +32,7 @@ class Cart extends Migration
             $table->string('sku');
             $table->string('name');
             $table->float('unit_price', 9, 2);
-            $table->integer('quantity');
+            $table->float('quantity', 9, 2);
 
             $table->text('details');
